MEGA65: DMA debug fix/optimization #198
Small fix for more sane debug output when DO_DEBUG_DMA is defined for
some extra DMA information. Nothing here which should affect an average
user, only in a very unlikely scenario:

One limitation of DMA emulation in Xemu: enhanced mode option list is
read in a blockling fashion, thus an "unended" list will stall the
emulator (though it's unlikely not to meet a zero byte sooner or later
in the memory). To avoid this (and also warn the user about something
fishy is going on) there is a warning window in that case - as the
number of enhanced mode options are limited, it's simply insane to have
a very large option list, thus it's a certain sign of some bug in the
software which produces this: better the user to know.

+++ targets/mega65/dma65.h
@@ -19,7 +19,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA */
 #ifndef XEMU_MEGA65_DMA_H_INCLUDED
 #define XEMU_MEGA65_DMA_H_INCLUDED
 
-extern int   in_dma;
+extern Uint8 in_dma;
 
 extern void  dma_write_reg                ( int addr, Uint8 data );
 extern Uint8 dma_read_reg                 ( int reg );
